// frontend/src/app/models/enum/SubmissionQueryParamEnum.ts
export enum SubmissionQueryParamEnum {
    PAGE = 'page',
    SIZE = 'size',
    SEARCH = 'search',
    SORT_BY = 'sortBy',
    SORT_DIRECTION = 'sortDir',

    ASSIGNMENT_ID = 'assignmentId',
    USER_ID = 'userId',
    STATUS = 'status',
    SUBMITTED_AFTER = 'submittedAfter',
    SUBMITTED_BEFORE = 'submittedBefore'
}